removing (DijkstraBase) duplicated code
