+++ functions/DijkstraBase.py
@@ -3,14 +3,14 @@ from pgRoutingLayer import pgRoutingLayer_utils as Utils
 from .FunctionBase import FunctionBase
 from psycopg2 import sql
 
-class Function(FunctionBase):
+class DijkstraBase(FunctionBase):
 
-    minPGRversion = 2.5
+    minPGRversion = 2.1
 
     @classmethod
     def getName(self):
         ''' returns Function name. '''
-        return 'pgr_bdDijkstra'
+        return 'pgr_dijkstra'
 
     @classmethod
     def getControlNames(self, version):
@@ -29,12 +29,12 @@ class Function(FunctionBase):
 
     @classmethod
     def getQuery(self, args):
-        ''' returns the sql query in required signature format of pgr_bdDijkstra '''
+        ''' returns the sql query in required signature format of pgr_dijkstra '''
         return sql.SQL("""
             SELECT seq, '(' || start_vid || ',' || end_vid || ')' AS path_name,
                 path_seq AS _path_seq, start_vid AS _start_vid, end_vid AS _end_vid,
                 node AS _node, edge AS _edge, cost AS _cost, lead(agg_cost) over() AS _agg_cost
-            FROM pgr_bdDijkstra('
+            FROM pgr_dijkstra('
                 {innerQuery}
                 ',
                 {source_ids}, {target_ids}, {directed})
@@ -45,7 +45,6 @@ class Function(FunctionBase):
 
     def getExportMergeQuery(self, args):
         return self.getExportManySourceManyTargetMergeQuery(args)
-
 
     def draw(self, rows, con, args, geomType, canvasItemList, mapCanvas):
         ''' draw the result '''
